+++ crates/software-architecture/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod software_architecture;
-pub mod software_rpc;